fixes 7

+++ sbin/pack.sh
@@ -36,6 +36,7 @@ if [[ $1 == --help || $1 == help || $HELP == 1 ]]; then
 
 		ARTDIR=dir          Directory in which packages are created (default: bin/artifacts)
 
+		JUST_PRINT=1        Only print package names, do not generate
 		VERBOSE=1           Print commands
 		HELP=1              Show help
 
@@ -44,14 +45,6 @@ if [[ $1 == --help || $1 == help || $HELP == 1 ]]; then
 fi
 
 #----------------------------------------------------------------------------------------------
-
-RAMP=${RAMP:-1}
-DEPS=${DEPS:-1}
-SYM=${SYM:-0}
-
-[[ -z $ARTDIR ]] && ARTDIR=bin/artifacts
-mkdir -p $ARTDIR $ARTDIR/snapshots
-ARTDIR=$(cd $ARTDIR && pwd)
 
 # RLEC naming conventions
 
@@ -66,16 +59,34 @@ OSNICK=$($READIES/bin/platform --osnick)
 [[ $OSNICK == xenial ]]  && OSNICK=ubuntu16.04
 [[ $OSNICK == bionic ]]  && OSNICK=ubuntu18.04
 [[ $OSNICK == focal ]]   && OSNICK=ubuntu20.04
-[[ $OSNICK == jammy ]] && OSNICK=ubuntu22.04
+[[ $OSNICK == jammy ]]   && OSNICK=ubuntu22.04
 [[ $OSNICK == centos7 ]] && OSNICK=rhel7
 [[ $OSNICK == centos8 ]] && OSNICK=rhel8
+[[ $OSNICK == ol8 ]]     && OSNICK=rhel8
 [[ $OSNICK == rocky8 ]]  && OSNICK=rhel8
+
+[[ $OSNICK == bigsur ]]  && OSNICK=catalina
+
+#----------------------------------------------------------------------------------------------
+
+if [[ -z $MODULE || ! -f $MODULE ]]; then
+	eprint "pack: MODULE is not defined or does not refer to a file"
+	exit 1
+fi
+
+RAMP=${RAMP:-1}
+DEPS=${DEPS:-0}
+SYM=${SYM:-0}
+
+[[ -z $ARTDIR ]] && ARTDIR=bin/artifacts
+mkdir -p $ARTDIR $ARTDIR/snapshots
+ARTDIR=$(cd $ARTDIR && pwd)
 
 export PRODUCT=redisbloom
 export PRODUCT_LIB=$PRODUCT.so
 export DEPNAMES=""
 
-export PACKAGE_NAME=${PACKAGE_NAME:-${PRODUCT}}
+export PACKAGE_NAME=redisbloom
 
 RAMP_CMD="python3 -m RAMP.ramp"
 
@@ -91,8 +102,9 @@ pack_ramp() {
 	
 	local fq_package=$stem.${verspec}.zip
 
+	[[ ! -d $ARTDIR ]] && mkdir -p $ARTDIR
+
 	local packfile="$ARTDIR/$fq_package"
-	local product_so="$MODULE"
 
 	local xtx_vars=""
 	local dep_fname=${PACKAGE_NAME}.${platform}.${verspec}.tgz
@@ -109,7 +121,7 @@ pack_ramp() {
 		$ROOT/$rampfile > /tmp/ramp.yml
 	rm -f /tmp/ramp.fname $packfile
 	$RAMP_CMD pack -m /tmp/ramp.yml --packname-file /tmp/ramp.fname --verbose --debug \
-		-o $packfile $product_so >/tmp/ramp.err 2>&1 || true
+		-o $packfile $MODULE >/tmp/ramp.err 2>&1 || true
 	if [[ ! -e $packfile ]]; then
 		eprint "Error generating RAMP file:"
 		>&2 cat /tmp/ramp.err
@@ -122,6 +134,8 @@ pack_ramp() {
 		ln -sf ../$fq_package $snap_package
 	fi
 
+	local packname=`cat /tmp/ramp.fname`
+	echo "Created $packname"
 	cd $ROOT
 }
 
@@ -131,8 +145,8 @@ pack_deps() {
 	local dep="$1"
 
 	local platform="$OS-$OSNICK-$ARCH"
+	local stem=${PACKAGE_NAME}.${dep}.${platform}
 	local verspec=${SEMVER}${_VARIANT}
-	local stem=${PACKAGE_NAME}-${dep}.${platform}
 
 	local depdir=$(cat $ARTDIR/$dep.dir)
 
@@ -143,12 +157,12 @@ pack_deps() {
 	{ cd $depdir ;\
 	  cat $ARTDIR/$dep.files | \
 	  xargs tar -c --sort=name --owner=root:0 --group=root:0 --mtime='UTC 1970-01-01' \
-		--transform "s,^,$dep_prefix_dir," 2>> /tmp/pack.err | \
+		--transform "s,^,$dep_prefix_dir," 2> /tmp/pack.err | \
 	  gzip -n - > $tar_path ; E=$?; } || true
 	rm -f $ARTDIR/$dep.prefix $ARTDIR/$dep.files $ARTDIR/$dep.dir
 
 	cd $ROOT
-	if [[ $E != 0 ]]; then
+	if [[ $E != 0 || -s /tmp/pack.err ]]; then
 		eprint "Error creating $tar_path:"
 		cat /tmp/pack.err >&2
 		exit 1
@@ -156,7 +170,7 @@ pack_deps() {
 	sha256sum $tar_path | awk '{print $1}' > $tar_path.sha256
 
 	cd $ARTDIR/snapshots
-	if [[ ! -z $BRANCH ]]; then
+	if [[ -n $BRANCH ]]; then
 		local snap_dep=$stem.${BRANCH}${_VARIANT}.tgz
 		ln -sf ../$fq_dep $snap_dep
 		ln -sf ../$fq_dep.sha256 $snap_dep.sha256
@@ -168,9 +182,10 @@ pack_deps() {
 #----------------------------------------------------------------------------------------------
 
 prepare_symbols_dep() {
+	if [[ ! -f $MODULE.debug ]]; then return 0; fi
 	echo "Preparing debug symbols dependencies ..."
 	echo $(cd "$(dirname $MODULE)" && pwd) > $ARTDIR/debug.dir
-	echo $PRODUCT.so.debug > $ARTDIR/debug.files
+	echo $MODULE.debug > $ARTDIR/debug.files
 	echo "" > $ARTDIR/debug.prefix
 	pack_deps debug
 	echo "Done."
@@ -204,6 +219,22 @@ if [[ $WITH_GITSHA == 1 ]]; then
 fi
 export BRANCH
 
+#----------------------------------------------------------------------------------------------
+
+if [[ $JUST_PRINT == 1 ]]; then
+	if [[ $RAMP == 1 ]]; then
+		echo "${PACKAGE_NAME}.${OS}-${OSNICK}-${ARCH}.${SEMVER}${VARIANT}.zip"
+	fi
+	if [[ $DEPS == 1 ]]; then
+		for dep in $DEPNAMES; do
+			echo "${PACKAGE_NAME}.${dep}.${OS}-${OSNICK}-${ARCH}.${SEMVER}${VARIANT}.tgz"
+		done
+	fi
+	exit 0
+fi
+
+#----------------------------------------------------------------------------------------------
+
 if [[ $DEPS == 1 ]]; then
 	echo "Building dependencies ..."
 
@@ -217,7 +248,7 @@ fi
 
 if [[ $RAMP == 1 ]]; then
 	if ! command -v redis-server > /dev/null; then
-		eprint "$PROGNAME: Cannot find redis-server. Aborting."
+		eprint "Cannot find redis-server. Aborting."
 		exit 1
 	fi
